+++ packages/components/evm/eslint.config.mjs
@@ -3,7 +3,7 @@ import baseConfig from '../../../eslint.config.mjs';
 export default [
   ...baseConfig,
   {
-    files: ['**/quick-node.provider.ts'],
+    files: ['**/quick-node.provider.ts', '**/rate-limiter.ts'],
     rules: {
       'no-restricted-syntax': 'off',
     },